MOT-1994 - libfranka server and client example (#2)
This PR adds a mock server and client example. The client iterates through all the libfranka calls used in the controls stack and the server responds appropriately. This will be a good check to use when we migrate to a newer version in the future.

The PR also adds some developer tooling:
* tasks.json for compile with Ctrl+Shift+B
* Makefile for command shortcuts
* fmt and magic_enum

Co-authored-by: Andreas Kuhner <andreas.kuhner@franka.de>